Improve server readability
